+++ src/main/java/com/mes/mesBackend/service/MeasureService.java
@@ -12,7 +12,7 @@ public interface MeasureService {
     // 계측기 단일 조회
     MeasureResponse getMeasure(Long id) throws NotFoundException;
     // 계측기 페이징 조회 검색조건: 검색조건: GAUGE유형, 검교정대상(월)
-    Page<MeasureResponse> getMeasures(Long gaugeId, int month, Pageable pageable);
+    Page<MeasureResponse> getMeasures(Long gaugeId, Long month, Pageable pageable);
     // 계측기 수정
     MeasureResponse updateMeasure(Long id, MeasureRequest measureRequest) throws NotFoundException;
     // 계측기 삭제
